requirements.txt
diff-match-patch==20121119
Django==1.9.6
django-grappelli==2.8.1
django-import-export==0.4.5
django-rest==0.0.1
django-tinymce==2.3.0
djangorestframework==3.3.3
Pillow==3.2.0
psycopg2==2.6.1
sorl-thumbnail==12.3
tablib==0.11.2
